progress

+++ papi.h
@@ -31,6 +31,7 @@
 #include "pcompat.h"
 #include "pcompiler.h"
 #include <string.h>
+#include <stdint.h>
 
 #define PARAM_STR   0
 #define PARAM_NUM   1
@@ -73,8 +74,8 @@ typedef struct _binresult{
   };
 } binresult;
 
-#define P_STR(name, val) {PARAM_STR, strlen(name), strlen(val), (name), {(uint64_t)(val)}}
-#define P_LSTR(name, val, len) {PARAM_STR, strlen(name), (len), (name), {(uint64_t)(val)}}
+#define P_STR(name, val) {PARAM_STR, strlen(name), strlen(val), (name), {(uint64_t)((uintptr_t)(val))}}
+#define P_LSTR(name, val, len) {PARAM_STR, strlen(name), (len), (name), {(uint64_t)((uintptr_t)(val))}}
 #define P_NUM(name, val) {PARAM_NUM, strlen(name), 0, (name), {(val)}}
 #define P_BOOL(name, val) {PARAM_BOOL, strlen(name), 0, (name), {(val)?1:0}}
